refactor(tests): move env_vars fixture to conftest.py
